Move domain and data tests to the packages of the classes under test
